requirements.txt
fastapi==0.115.11
mangum==0.19.0
pydantic==2.10.6
pydantic-settings==2.8.1
boto3==1.37.15
requests==2.32.3
python-dotenv==1.0.1
uvicorn==0.34.0 